UI audit: ZOC toggle, fire/melee previews, transport, advance, stepper, hover warnings
Implement the full UI audit gap list (P1-P7):

- P1: ZOC overlay toggle button (overlay_toggles_ui) at top-right
- P2: Fire preview now shows range/hex, range band, night indicator,
  and LOS status with blocking hex/hexside and § citation
- P3: Melee preview now shows per-modifier breakdown with § citations,
  CRT row labels, and outcome bands for both attacker and defender
- P4a/b: Special actions panel for Zariba construction (§5.3) and
  Royal Engineers demolition (§6.53)
- P4c: Friendlies transport UI for Load/Cross/Disembark (§5.21)
- P5: Advance-after-combat prompt shows vacated hex count when unit
  can advance after fire/melee
- P7a: Phase stepper shows ✓ for completed phases
- P7b: Tribe-mixing warning in hover tooltip (§5.52)
- P7c: Road bonus indicator in hover tooltip when road reduces cost

Also: fix 4 compiler warnings (unreachable pattern, unused mut, dead
code), fix all clippy warnings across workspace.

+++ omdurman-app/src/turn_track_ui.rs
@@ -148,7 +148,7 @@ pub(crate) fn turn_track_labels(
         let n_cols = if row == 2 { 4 } else { 9u8 };
         for col in 0..n_cols {
             let idx = row * 9 + col;
-            let turn_num = (idx + 1) as u8;
+            let turn_num = idx + 1;
             let label = omdurman_rules::turn_track::TurnLabel::from_turn(turn_num);
 
             let cx_px = match row {
